+++ docs/ChangeLog.md
@@ -103,6 +103,7 @@
 - edit: render the final prompt before updating command history (motivated by pallaswept) `#D2207` 911a4051
   - debug: support `bleopt debug_profiler_opts=tree` `#D2216` xxxxxxxx
 - keymap/vi: fix <kbd>C-w</kbd> not saving the word into kill ring `#D2208` aa7ca45d
+- edit: fix standard streams in `EXIT` trap with `ble/widget/exit` `#D2217` xxxxxxxx
 
 ## Compatibility